Drop 'use autodie' everywhere; handle errors explicitly.
The 'autodie' pragma breaks many expectations for error handling and
fails to emulate parser quirks in functions like system and exec. Much
of that behavior is documented in the last paragraph here, and also in
the "Gotchas" section below:

    https://perldoc.perl.org/autodie#FUNCTION-SPECIFIC-NOTES

Thanks to Grinnz from IRC:#perl-help for these pointers!

Gbp-Dch: ignore

+++ lib/Test/Lintian/Prepare.pm
@@ -37,7 +37,6 @@ tests are run. To do so, they use the specifications in the test set.
 use v5.20;
 use warnings;
 use utf8;
-use autodie;
 
 use Exporter qw(import);
 
@@ -155,7 +154,10 @@ sub prepare {
 
     # delete old test scripts
     my @oldrunners = File::Find::Rule->file->name('*.t')->in($sourcepath);
-    unlink(@oldrunners);
+    if (@oldrunners) {
+        unlink(@oldrunners)
+          or die "Cannot unlink @oldrunners";
+    }
 
     my $skeletonname = $desc->unfolded_value('Skeleton');
     if (length $skeletonname) {
@@ -346,7 +348,10 @@ sub filleval {
 
     # delete old test scripts
     my @oldrunners = File::Find::Rule->file->name('*.t')->in($evalpath);
-    unlink(@oldrunners);
+    if (@oldrunners) {
+        unlink(@oldrunners)
+          or die "Cannot unlink @oldrunners";
+    }
 
     $testcase->store('Skeleton', $desc->value('Skeleton'))
       unless $testcase->declares('Skeleton');
